stop using SWAP_INT_PROC [fixes #3]
It is not translated by windows extender.
It actually cannot be reliably translated because it can't return
the prev handler pointer in a way that is usable for the subsequent
swap. It can only return the prev PM handler or NULL, which is what
dosemu does, but it can't return prev RM handler.

Thanks to @Baron-von-Riedesel for nailing the problem.